+++ src/modules/Form/Form.module.less
@@ -1,82 +1,7 @@
-
-.root {
-
-    /* Change the white to any color */
-    input:-webkit-autofill,
-    input:-webkit-autofill:hover, 
-    input:-webkit-autofill:focus, 
-    input:-webkit-autofill:active
-    {
-      transition: background-color 5000s;
-      -webkit-text-fill-color: #fff !important;
-    }
-
-    box-sizing: border-box;
-
-    .container, .header, .footer {
-        margin: 0;
-        padding: 0;
-        box-sizing: content-box;
-    }
+.wrap {
+  text-align: left;
 }
 
-.button {
-  -webkit-appearance: none;
-  display: inline-block;
-  text-align: center;
-  vertical-align: middle;
-  cursor: pointer;
-  border: none;
-  position: relative;
-  min-height: 50px;
-  user-select: none;
-  &:hover {
-    outline: 0;
-  }
-
-  &:focus {
-    outline: 0;
-  }
-
-  &:active {
-    outline: 0;
-  }
-
-  &:disabled {
-    pointer-events: none;
-  }
-}
-
-
-.root :global {
-  .MuiGrid-container {
-      margin: 0;
-      width: 100%;
-  }
-
-  input:-internal-autofill-selected {
-    background-color: transparent !important;
-  }
-
-  ::-webkit-calendar-picker-indicator{
-    opacity: 0.3;
-  }
-
-  .MuiScopedCssBaseline-root {
-      background-color:transparent;
-      color: inherit;
-  }
-
-  .MuiInputBase-root, .MuiScopedCssBaseline-root, .MuiSvgIcon-root, .MuiRadio-root, .MuiCheckbox-root, .MuiFormLabel-root {
-      color: inherit;
-  }
-
-  .MuiSvgIcon-root, .MuiIconButton-root, .MuiTypography-body1, .MuiInputBase-root {
-    font-size: inherit;
-  }
-
-  .MuiFormLabel-root.Mui-focused {
-    color: inherit;
-  }
-  
+.form {
+  width: 100%;
 }
