+++ src/modules/FormBak/index.ts
@@ -1,0 +1,2 @@
+import Form from './Form';
+export default Form;
